+++ gallery/animate-abeamer/js/main.js
@@ -26,12 +26,12 @@ $(window).on("load", function () {
                 {
                     prop: 'opacity',
                     duration: '0.5s',
+                    advance: true,
                 },
                 {
                     prop: 'transform',
                     valueFormat: 'rotate(%fdeg)',
                     duration: '1s',
-                    position: '+0.5s',
                     value: 100,
                 }
             ],
@@ -59,12 +59,12 @@ $(window).on("load", function () {
                 {
                     prop: 'opacity',
                     duration: '0.5s',
+                    advance: true,
                 },
                 {
                     prop: 'transform',
                     valueFormat: 'rotate(%fdeg)',
                     duration: '1s',
-                    position: '+0.5s',
                     value: 100,
                 }
             ],
@@ -122,12 +122,12 @@ $(window).on("load", function () {
                 {
                     prop: 'left',
                     duration: '1s',
+                    advance: true,
                     easing: 'easeInCubic',
                     value: 52,
                 },
                 {
                     prop: 'left',
-                    position: '+1s',
                     duration: '2s',
                     oscillator: {
                         handler: 'damped',